style: 프로필 영역 스타일 수정
